refactor: section 관련 기능 line 패키지로 이동
