+++ src/main/java/subway/line/Line.java
@@ -1,7 +1,5 @@
 package subway.line;
 
-import subway.section.Section;
-import subway.section.Sections;
 import subway.station.Station;
 
 import javax.persistence.*;
