+++ src/main/java/subway/line/LineService.java
@@ -61,4 +61,37 @@ public class LineService {
     public void deleteLine(Long id) {
         lineRepository.deleteById(id);
     }
+
+    @Transactional
+    public Long saveSection(Long lineId, SectionRequest request) {
+        Line line = lineRepository.findById(lineId)
+                .orElseThrow(() -> new IllegalStateException(String.format("지하철 노선을 찾을 수 없습니다. (id: %d)", lineId)));
+
+        Station upStation = stationRepository.findById(request.getUpStationId())
+                .orElseThrow(() -> new IllegalStateException(String.format("노선의 하행종점역을 찾을 수 없습니다. (downStationId: %d)", request.getUpStationId())));
+
+        if (!line.getDownStationId().equals(upStation.getId())) {
+            throw new IllegalStateException(String.format("새로운 구간의 상행역은 해당 노선에 등록되어있는 하행 종점역이 아닙니다. (구간의 upStationId: %d)", request.getUpStationId()));
+        }
+
+        Station downStation = stationRepository.findById(request.getDownStationId())
+                .orElseThrow(() -> new IllegalStateException(String.format("새로운 하행종점역을 찾을 수 없습니다. (downStationId: %d)", request.getDownStationId())));
+
+        if (line.contains(downStation)) {
+            throw new IllegalStateException(String.format("새로운 구간의 하행역은 해당 노선에 등록되어있는 역일 수 없습니다. (downStationId: %d)", downStation.getId()));
+        }
+
+        Section section = new Section(line, upStation, downStation, request.getDistance());
+        line.addTerminalSection(section);
+
+        return section.getId();
+    }
+
+    @Transactional
+    public void deleteSection(Long lineId, Long stationId) {
+        Line line = lineRepository.findById(lineId)
+                .orElseThrow(() -> new IllegalStateException(String.format("지하철 노선을 찾을 수 없습니다. (id: %d)", lineId)));
+
+        line.deleteSectionByDownStationId(stationId);
+    }
 }
