--- src/main/java/subway/section/SectionRepository.java
@@ -1,7 +1,0 @@
-package subway.section;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SectionRepository extends JpaRepository<Section, Long> {
-    void deleteAllByLineId(Long lineId);
-}
